Updated getAccount() to return the entire account object instead of just the account name.
Calling login() with a code will bypass the token store and retrieve fresh tokens.

Fixed bug where the redirect_uri contained undefined when doing manual login() with a code.

Updated the readme with some clarifications and fixes.

Added a couple new more unit tests.

Added a refresh prop to getAccount() (defaults to true) that will auto-refresh an account's access token if it's expired and the refresh token is still valid.